+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Added
 - Version wildcards support for single '*' to match all.
+- Added support to run a component server and handle a single request
+  where the payload is send from the CLI in a JSON file (#72).
 
 ### Changed
 - The wildcard ('*') in the last version part now matches any character.